Big code refactor in scripts/master.js
